V10
This is the fixed motor target, speed is still iffy and we need to do the test's with this code to make up V11. 

This includes all code uncommented but we will be commenting it out so it works how we need it to. 

Our test will be: 
How far we move forward when moving at a speed for a time
How much we turn at a speed for a time
Where the servos go to with the current open const int's

If the testing works how we need, we will adjust our code to work how we need, and make it function in an order we want it too.